feat: Add divide_by_size parameter to create_product_revenue_bar_chart
This commit adds a new parameter, divide_by_size, to the create_product_revenue_bar_chart function in page7.py. When set to True, the function will display the product name along with its size in the y-axis labels of the bar chart. When set to False, only the product name will be displayed. This enhancement provides flexibility in customizing the chart based on the user's preference.