feat: Update hoverinfo and textinfo in create_customers_pie_chart function
This commit updates the hoverinfo and textinfo parameters in the create_customers_pie_chart function in the page6.py file. The hoverinfo parameter is changed from 'label+percent' to 'value+label+percent', and the textinfo parameter is changed from 'value+percent' to 'value+percent'. This change improves the display of information in the customers pie chart and enhances the user experience by providing more detailed data.